requirements.txt
asttokens==2.4.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.47.2
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.3
packaging==23.2
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
zipp==3.17.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pillow==10.2.0
Werkzeug==3.0.1
zipp==3.17.0
